+++ internal/adventuria/actions/custom/update_review/action.go
@@ -38,9 +38,9 @@ func (u *UpdateReview) CanDo(_ context.Context, _ *model.Events, _ *model.Player
 }
 
 type Request struct {
-	ActionID string  `json:"action_id" form:"action_id"`
-	Comment  *string `json:"comment" form:"comment"`
-	Score    *int    `json:"score" form:"score"`
+	ActionID string   `json:"action_id" form:"action_id"`
+	Comment  *string  `json:"comment" form:"comment"`
+	Score    *float64 `json:"score" form:"score"`
 }
 
 func (u *UpdateReview) Do(ctx context.Context, _ *model.Events, _ *model.Player, actionReq model.ActionRequest) (any, error) {
